The following commit will: 1. change the backend url by removing slash 2. Change log entry value formatting 3. Update the doc of xml seriallize in api helpers 4. DISCARD DECIMALS in dollars in flight

+++ src/covid19ImpactEstimator.py
@@ -88,7 +88,7 @@ def covid19ImpactEstimator(impact, timeToElapseInDays, **data):
 	cFICUBRT = int(iBRT * 0.05)
 	cFVBRT = int(iBRT * 0.02)
 
-	dollarsInFlight = int((iBRT * aDIIU * aDIP) / timeToElapseInDays)
+	dollarsInFlight = int(int((iBRT * aDIIU * aDIP)) / timeToElapseInDays)
 
 	return {
 		"currentlyInfected": cI,
